Fixed a bug where adding an offset and length resulting in integer overflow was incorrectly passing a bounds check. Added unit test for IccReader (related to above issue.) Created utility class ExtractAppSegmentBytesToFileUtility for writing specific JPG segments to files, and reading them again into byte[] for convenience in unit testing (and keeping source code file size down.)
git-svn-id: https://metadata-extractor.googlecode.com/svn/trunk@106 7bd79e3c-ab28-1fa5-d007-79fa9fde61af

+++ Source/com/drew/lang/BufferReader.java
@@ -285,7 +285,7 @@ public class BufferReader
 
     private void CheckBounds(final int index, final int bytesRequested) throws BufferBoundsException
     {
-        if (bytesRequested < 0 || index < 0 || index + bytesRequested - 1 >= _buffer.length)
+        if (bytesRequested < 0 || index < 0 || (long)index + (long)bytesRequested - 1L >= (long)_buffer.length)
             throw new BufferBoundsException(_buffer, index, bytesRequested);
     }
 }
